backend/requirements.txt
fastapi
uvicorn
transformers
pinecone-client==3.1.0
openai==0.27.7
python-dotenv
langchain==0.1.1
langchain-community==0.0.13
langchain_openai